Added a new ALT_NICK option for reconnect issues
Reconnecting adds a random number to the nick, which sucks if you have
a password in your rc file as it will try to wait for +e to be set
before it autojoins channels (otherwise your cloak will not be set
before you join, for example). This way you can set an alt nick which
you have joined to your main nick and it will reconnect with that.

I haven't tested it yet - I will have to wait until the next time I get
disconnected, but hopefully it should work.